rename "sugoi.js" to "sugois.js" for consistency with command name
